feat: my 안으로 이동 (#27)
