+++ src/pages/my/my.css.ts
@@ -1,0 +1,6 @@
+import { style } from '@vanilla-extract/css';
+
+export const container = style({
+  height: '100dvh',
+  background: 'linear-gradient(355deg, #FFF 48.23%, #BFD9FE 97.05%), #FFF',
+});
