+++ src/pages/reviewNotes/ReviewNotes.tsx
@@ -1,21 +1,5 @@
-import Chip from '@components/chip/Chip';
-import { CHIP_LIST } from '@components/chip/chipData';
-
-import * as styles from './reviewNotes.css';
-
 const ReviewNotes = () => {
-  return (
-    <div className={styles.container}>
-      {CHIP_LIST.map((chip, idx) => (
-        <Chip
-          key={idx}
-          icon={chip.icon}
-          label={chip.label}
-          background={chip.background}
-        />
-      ))}
-    </div>
-  );
+  return <div>오답노트목록</div>;
 };
 
 export default ReviewNotes;
