--- src/pages/reviewNotes/reviewNotes.css.ts
@@ -1,6 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const container = style({
-  height: '100dvh',
-  background: 'linear-gradient(355deg, #FFF 48.23%, #BFD9FE 97.05%), #FFF',
-});
